Refactor ID and Widget models definition
