+++ src/model/id.ts
@@ -1,0 +1,7 @@
+export type ID = string;
+
+export function generateId(): ID {
+  const timestamp = Date.now().toString(36);
+  const random = Math.random().toString(36).slice(2, 8);
+  return `${timestamp}${random}`;
+}
