+++ src/model/widget_settings.ts
