+++ src/store/workfield.ts
@@ -1,35 +1,16 @@
 import { defineStore } from "pinia";
 import { type Component as AsyncComponent } from "vue";
-import { WidgetSettings } from "./widget_settings";
-
-interface Input {
-  data: Number | String | Object | Array<Object>;
-}
-
-interface Output {
-  data: Number | String | Object | Array<Object>;
-}
-
-function generateWidgetId(): string {
-  const timestamp = Date.now().toString(36);
-  const random = Math.random().toString(36).slice(2, 8);
-  return `${timestamp}${random}`;
-}
+import { WidgetSettings } from "../model/widget_settings";
+import { type Widget } from "../model/widget";
+import { generateId, type ID } from "../model/id";
 
 interface WidgetPreview {
   widgetType: AsyncComponent;
   settings: WidgetSettings;
 }
 
-export interface Widget {
-  widgetType: AsyncComponent;
-  settings: WidgetSettings;
-  input?: Input;
-  output?: Output;
-}
-
 export interface WorkfieldState {
-  widgets: Record<string, Widget>;
+  widgets: Record<ID, Widget>;
 }
 
 export const useWorkfieldStore = defineStore("workfield", {
@@ -40,7 +21,7 @@ export const useWorkfieldStore = defineStore("workfield", {
   },
   actions: {
     addWidget({ settings, widgetType }: WidgetPreview) {
-      this.widgets[generateWidgetId()] = { settings, widgetType };
+      this.widgets[generateId()] = { settings, widgetType };
     },
     updateWidget(id: string, settings: WidgetSettings) {
       this.widgets[id].settings = settings;
